fix(WebAPI): Fixed the endpoint for hiding notifications
