+++ src/Application/Notifications/Hide/SetNotificationVisibilityEndpoint.cs
@@ -1,0 +1,55 @@
+using Application.Contracts;
+using Data.Contracts;
+using FastEndpoints;
+using FluentValidation;
+using Microsoft.AspNetCore.Http;
+using Microsoft.EntityFrameworkCore;
+
+namespace PlexRipper.Application;
+
+public record SetNotificationVisibilityEndpointRequest()
+{
+    public int Id { get; init; }
+
+    public bool Hidden { get; init; } = true;
+}
+
+public class SetNotificationVisibilityEndpointRequestValidator : Validator<SetNotificationVisibilityEndpointRequest>
+{
+    public SetNotificationVisibilityEndpointRequestValidator()
+    {
+        RuleFor(x => x.Id).GreaterThan(0);
+        RuleFor(x => x.Hidden).Equal(true);
+    }
+}
+
+public class SetNotificationVisibilityEndpoint : BaseEndpoint<SetNotificationVisibilityEndpointRequest, ResultDTO>
+{
+    private readonly IPlexRipperDbContext _dbContext;
+
+    public override string EndpointPath => ApiRoutes.NotificationController;
+
+    public SetNotificationVisibilityEndpoint(IPlexRipperDbContext dbContext)
+    {
+        _dbContext = dbContext;
+    }
+
+    public override void Configure()
+    {
+        Patch(EndpointPath);
+        AllowAnonymous();
+        Description(x =>
+            x.Produces(StatusCodes.Status200OK, typeof(ResultDTO))
+                .Produces(StatusCodes.Status500InternalServerError, typeof(ResultDTO))
+        );
+    }
+
+    public override async Task HandleAsync(SetNotificationVisibilityEndpointRequest req, CancellationToken ct)
+    {
+        await _dbContext
+            .Notifications.Where(x => x.Id == req.Id)
+            .ExecuteUpdateAsync(x => x.SetProperty(y => y.Hidden, req.Hidden), ct);
+
+        await SendFluentResult(Result.Ok(), ct);
+    }
+}
